Created setPosition, gyro_loop_period in Drivetrain
setPosition will set the worldPositions and worldAngle_rad given the inputs, and will convert degrees from the user into rads for the localizer. gyro_loop_period will reference the gyro position every few loops (as many loops as specified by the user. This is designed to subtly increase the effectiveness of the position tracking (of the heading). WIP, needs to interact with offset